+++ engine/app/components/citizens_advice_components/footer.rb
@@ -2,9 +2,10 @@
 
 module CitizensAdviceComponents
   class Footer < Base
-    attr_reader :homepage_url, :feedback_url
+    attr_reader :feedback_url
 
     renders_one :feedback_link, "FooterFeedbackLink"
+    renders_one :logo, "FooterLogo"
 
     renders_one :legal_summary, lambda { |text|
       text.presence || t("citizens_advice_components.footer.legal_summary")
@@ -12,9 +13,8 @@ module CitizensAdviceComponents
 
     renders_many :columns, "FooterColumn"
 
-    def initialize(homepage_url: nil, feedback_url: nil)
+    def initialize(feedback_url: nil)
       super
-      @homepage_url = homepage_url || "/"
       @feedback_url = feedback_url.to_s
 
       feedback_url_deprecation
@@ -62,6 +62,22 @@ module CitizensAdviceComponents
       return if legal_summary
 
       t("citizens_advice_components.footer.legal_summary")
+    end
+
+    class FooterLogo < Base
+      attr_reader :title, :url
+
+      def initialize(title: nil, url: "/")
+        super
+        @title = title
+        @url = url
+      end
+
+      def call
+        # Renders a block if provided to allow passing a custom logo SVG,
+        # otherwise renders the standard logo.
+        content.presence || link_to("", url, title: title || t("citizens_advice_components.footer.logo_title"), class: "cads-logo")
+      end
     end
 
     class FooterColumn < Base