+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+**Breaking changes**
+
+- Footer: Replaced the logo background with a `logo` slot. Removed the `homepage_url`.
+
 ## v5.5.0
 
 ### 5 September 2023